+++ docs/changes.md
@@ -2,6 +2,7 @@
 2022/04/01
   * 完善给edit、mledit和combo\_box增加了empty\_over状态（感谢雨欣提供补丁）
   * 完善URL解析并增加一些新的schema。
+  * 完善ubjson writer
 
 2022/03/31
   * 增加带统计功能的stream （感谢林福提供)